hosts: adding google-kukui!
Since it is a very small machine, I have to create new options to
disable some of the packages to be built for this and save some space as
well.
In the future, I might want to change it with something more obvious
s.a. "terminal packages" and "gui packages" or something like that.